# Buffer chunking and more