--- a/tests/testutils.py
+++ b/tests/testutils.py
@@ -23,4 +23,5 @@ def _determine_can_use_wgpu_lib():
     )
 
 
+
 can_use_wgpu_lib = _determine_can_use_wgpu_lib()