… docs/guide.rst
… pygfx/resources/_texture.py
… pygfx/resources/_utils.py
… tests/geometries/test_bounding_box.py
… tests/resources/test_buffer.py
… tests/resources/test_chunk_sizes.py
… tests/resources/test_texture.py