Closes #19, closes #600, closes #812, closes #714.

## Tasks

* [x] Implement chunked buffer uploads.
* [x] Implement chunked texture uploads.
* [x] Unit tests for buffer.
* [x] Unit tests for texture.
* [x] Perform benchmarks to compare performance.
* [x] See if we can reduce chunksize to 1/32th of size?
* [x] Check if max chunk size degrades performance by increased Python looping.
* [x] Implement  `update_indices()`.
* [x] Remove ambiguity between NHW and HWC (e.g. using explicit size).
* [x] Refactor how rgb textures are handled.
* [x] Docs.
* [x] Support for big endian. 

Follow up:
* #828

## Benchmarks buffer

Before:
```
      upload_buffer_full_naive (20x) - cpu: 21.61
  upload_buffer_full_optimized (20x) - cpu: 21.69
    upload_buffer_full_noncont (20x) - cpu: 95.19
            upload_buffer_half (20x) - cpu: 10.25
    upload_buffer_two_quarters (20x) - cpu: 19.99
            upload_100_buffers (20x) - cpu: 21.16
```
After:
```
      upload_buffer_full_naive (20x) - cpu: 21.45
  upload_buffer_full_optimized (20x) - cpu: 19.52
    upload_buffer_full_noncont (20x) - cpu: 52.75
            upload_buffer_half (20x) - cpu: 10.03
    upload_buffer_two_quarters (20x) - cpu: 10.21
            upload_100_buffers (20x) - cpu: 23.25
```

Observations:
* The new `set_data()` method avoids a datacopy, though this only makes it a wee bit faster.
* Performance for non-contiguous data is less than for contiguous data, as expected, but it has improved nearly 2x!
* The chunking makes that updating the first and last quarter is about as fast as updating half.
* Updating many buffers seems a wee bit slower. Will look into that.


## Benchmarks texture

*will update as things progress*

Before:
```
       upload_tex1d_full_naive (20x) - cpu:  0.11 ms
   upload_tex1d_full_optimized (20x) - cpu:  0.12 ms
     upload_tex1d_full_noncont (20x) - cpu:  0.31 ms
           upload_tex1d_half_x (20x) - cpu:  0.12 ms
   upload_tex1d_two_quarters_x (20x) - cpu:  0.14 ms
                      divider1 (20x) - cpu:  0.00 ms
       upload_tex2d_full_naive (20x) - cpu: 12.26 ms
   upload_tex2d_full_optimized (20x) - cpu: 13.37 ms
     upload_tex2d_full_noncont (20x) - cpu:171.32 ms
           upload_tex2d_half_x (20x) - cpu: 10.75 ms
           upload_tex2d_half_y (20x) - cpu:  6.35 ms
   upload_tex2d_two_quarters_x (20x) - cpu: 12.39 ms
   upload_tex2d_two_quarters_y (20x) - cpu:  6.39 ms
                      divider2 (20x) - cpu:  0.00 ms
       upload_tex3d_full_naive (20x) - cpu: 27.25 ms
   upload_tex3d_full_optimized (20x) - cpu: 25.88 ms
     upload_tex3d_full_noncont (20x) - cpu:131.90 ms
           upload_tex3d_half_x (20x) - cpu: 20.47 ms
           upload_tex3d_half_y (20x) - cpu: 19.09 ms
           upload_tex3d_half_z (20x) - cpu: 12.77 ms
   upload_tex3d_two_quarters_x (20x) - cpu: 38.93 ms
   upload_tex3d_two_quarters_y (20x) - cpu: 19.63 ms
   upload_tex3d_two_quarters_z (20x) - cpu: 13.27 ms
```

After:
```
         upload_tex1d_full_naive (20x) - cpu:  0.12 ms
   upload_tex1d_full_optimized (20x) - cpu:  0.10 ms
     upload_tex1d_full_noncont (20x) - cpu:  0.19 ms
        upload_tex1d_quarter_x (20x) - cpu:  0.10 ms
     upload_tex1d_two_eights_x (20x) - cpu:  0.14 ms
                      divider1 (20x) - cpu:  0.00 ms
       upload_tex2d_full_naive (20x) - cpu: 13.92 ms
   upload_tex2d_full_optimized (20x) - cpu: 13.40 ms
     upload_tex2d_full_noncont (20x) - cpu: 95.69 ms
        upload_tex2d_quarter_x (20x) - cpu: 14.40 ms
        upload_tex2d_quarter_y (20x) - cpu:  3.64 ms
     upload_tex2d_two_eights_x (20x) - cpu: 13.97 ms
     upload_tex2d_two_eights_y (20x) - cpu:  3.20 ms
    upload_tex2d_four_eights_x (20x) - cpu: 13.38 ms
    upload_tex2d_four_eights_y (20x) - cpu:  6.51 ms
  upload_tex2d_chunk_stripes_x (20x) - cpu: 12.83 ms
  upload_tex2d_chunk_stripes_y (20x) - cpu:  7.22 ms
          upload_tex2d_random8 (20x) - cpu:  0.48 ms
         upload_tex2d_random16 (20x) - cpu:  0.98 ms
         upload_tex2d_random32 (20x) - cpu:  1.58 ms
         upload_tex2d_random64 (20x) - cpu:  3.18 ms
        upload_tex2d_random128 (20x) - cpu:  6.22 ms
        upload_tex2d_random256 (20x) - cpu: 11.78 ms
        upload_tex2d_random512 (20x) - cpu: 12.92 ms
       upload_tex2d_random1024 (20x) - cpu: 12.24 ms
       upload_tex2d_random2048 (20x) - cpu: 11.90 ms
     upload_tex2d_100_textures (20x) - cpu: 19.03 ms
                      divider2 (20x) - cpu:  0.00 ms
       upload_tex3d_full_naive (20x) - cpu: 29.33 ms
   upload_tex3d_full_optimized (20x) - cpu: 27.38 ms
     upload_tex3d_full_noncont (20x) - cpu: 81.43 ms
        upload_tex3d_quarter_x (20x) - cpu: 25.59 ms
        upload_tex3d_quarter_y (20x) - cpu: 26.64 ms
        upload_tex3d_quarter_z (20x) - cpu:  6.69 ms
     upload_tex3d_two_eights_x (20x) - cpu: 25.09 ms
     upload_tex3d_two_eights_y (20x) - cpu: 26.88 ms
     upload_tex3d_two_eights_z (20x) - cpu:  7.01 ms
    upload_tex3d_four_eights_x (20x) - cpu: 25.50 ms
    upload_tex3d_four_eights_y (20x) - cpu: 24.74 ms
    upload_tex3d_four_eights_z (20x) - cpu: 14.78 ms
  upload_tex3d_chunk_stripes_x (20x) - cpu: 25.70 ms
  upload_tex3d_chunk_stripes_y (20x) - cpu: 25.10 ms
  upload_tex3d_chunk_stripes_z (20x) - cpu: 14.07 ms
          upload_tex3d_random8 (20x) - cpu:  0.62 ms
         upload_tex3d_random16 (20x) - cpu:  0.97 ms
         upload_tex3d_random32 (20x) - cpu:  1.98 ms
         upload_tex3d_random64 (20x) - cpu:  3.64 ms
        upload_tex3d_random128 (20x) - cpu:  7.27 ms
        upload_tex3d_random256 (20x) - cpu: 18.46 ms
        upload_tex3d_random512 (20x) - cpu: 24.61 ms
       upload_tex3d_random1024 (20x) - cpu: 24.29 ms
       upload_tex3d_random2048 (20x) - cpu: 24.25 ms
```

Observations:
* The new `set_data()` method avoids a datacopy, though this only makes it a wee bit faster.
* Performance for non-contiguous data has improved a lot!
* Other than that, uploads are generally slower :/  work to do.

## API changes

* New `buffer.set_data()` and `texture.set_data()` methods.
* New `buffer.update_full()` and `texture.update_full()` methods.
* New `buffer.update_indices()` and `texture.update_indices()` methods.
* New `force_contiguous` arg for `Buffer()` and `Texture()` for cases where data is updated often.
* New `chuck_size` arg for `Buffer()` and `Texture()` to override the chunk size for syncing data.
* Can set `nitems` when creating a `Buffer` to override the interpreted number of items from the shape of the array.
* Can set `size` when creating a `Texture` to override the interpreted size from the shape of the array.
* The `.mem` attribute of `Buffer` and `Texture` is renamed to `.view` and is an `np.ndarray` instead of a `memoryview`. 
* It is no longer possible to create zero-sized buffers or textures (wgpu crashes when attempting to use one).
* Buffers and textures support data in big endian format.