@almarklein 2024-06-27
Status: all the chunking stuff is implemented and tested. Next up is doing some benchmarking so I can tweak the implementation and chunk parameters for performance.