@panxinmiao 2024-06-28
When users explicitly specify the size of a Texture, can we disregard the shape of the data, treat it solely as binary data, and use the user-specified size as the Texture size (with appropriate checks)? I believe this approach can offer better flexibility for dynamic texture data updates.
﻿
Additionally, I think it is crucial for users to explicitly specify the size when initializing a Texture. Of course, if users only provide data without specifying the size, our API can attempt to infer the size from the data's shape (using the current `size_from_data()` method). However, this inference is not always accurate. We should issue a warning and advise users to avoid relying on this method because, fundamentally, we cannot determine the texture's size attribute solely from the shape.
﻿
If `len(shape) == 4`, it is most likely in NHWC layout, where the size is (W, H) and the number of channels is C, which is straightforward.
﻿
If `len(shape) == 3`, theoretically, it could be NHW (n grayscale images) or HWC (a single bitmap). We cannot determine this definitively. We can only assume it is HWC (as is the case most of the time) and issue a warning.
﻿
Moreover, not only can N and C be 1, but H and W can also be 1. In such cases, if any dimension of 1 is collapsed, the situation becomes even more complicated.