@almarklein 2024-07-02
The `dim` arg is meant to remove this ambiguity. It can indeed fix it for channels, but I realize now that it cannot really disambiguate for `N`, nor any collapsed dimensions. I think the latter is not that much of a problem, because the user can create a view that has singleton dimensions. But the former is trickier.  Thanks for bringing this up!