@panxinmiao 2024-07-04
I think that when users explicitly specify the size of a Texture, we should disregard other conditions and adhere to the user-specified size, adjusting the data to fit this size. This approach can provide a fallback solution for most edge cases. 
Only when the user does not specify a size should we attempt to infer the size automatically, issuing a warning if there is any ambiguity.