@almarklein 2024-07-05
> I think that when users explicitly specify the size of a Texture, we should disregard other conditions and adhere to the user-specified size, adjusting the data to fit this size.

Done! Also changed the internal memoryview to a ndarray instead, which is reshaped to the texture size, which simplifies the code in a few places.

> Even, we can force users to specify a size when create texture.

Note that `dim` is mandatory, and together with the shape, we can derive the texture size pretty well, if we simply assume its not a stack. This makes all stack-less use-case simpler.