@almarklein 2024-07-09
Created benchmarks for texture. For now, the chunking eum ... negatively affects performance. I'll try to find out whether this is just due to chunking too much, or due to the overhead of calculating the chunks.