@hmaarrfk 2024-07-23
What command do i run on my linux + gpu machines t report these benchmarking numbers that you have at the top?