@almarklein 2024-07-23
Run this: https://github.com/pygfx/pygfx-benchmarks/blob/main/benchmarks/bm_texture.py